fix: polish-pass bug hunt — YouTube vanish, /events auth, +more
Backend:
- youtube api_get now fails on non-2xx (403 quotaExceeded etc.) instead of
  deserializing the error body into an all-Option struct as an empty result,
  which silently dropped every YouTube channel from the rail with no log.
- twitch device-code poll bails on non-400/unparseable errors instead of
  spinning silently until the code expires.
- twitch lookup_channel_id_by_login passes login as an encoded query param and
  returns (id, display_name); wizard shows the real cased name.
- persist add_blocklist rejects empty vod_id (channel-sentinel collision).
Web:
- /events SSE now requires auth (X-Api-Key or session cookie) — was open.
- auth length-compare uses usize (u8 aliased lengths mod 256).
- logs tail matches .log suffix; safe_backup_name rejects dotfiles.
Frontend:
- SSE reconnects after a hard close (401 pre-login) + (re)connects on login.
- escape backup filenames, wizard message; validate VOD href scheme (no
  javascript:); await renderPipelines/Plugins/Recordings.
- de-scope .platform-icon/.status/.actions so channel-detail renders styled.

+++ crates/strivo-web/src/auth.rs
@@ -108,10 +108,12 @@ impl ApiKey {
     pub fn matches(&self, candidate: &str) -> bool {
         let a = self.0.as_bytes();
         let b = candidate.as_bytes();
-        let mut diff: u8 = (a.len() as u8) ^ (b.len() as u8);
+        // Use usize for the length diff: casting to u8 would alias lengths that
+        // are congruent mod 256 (e.g. 32 vs 288), defeating the length guard.
+        let mut diff: usize = a.len() ^ b.len();
         let n = a.len().min(b.len());
         for i in 0..n {
-            diff |= a[i] ^ b[i];
+            diff |= (a[i] ^ b[i]) as usize;
         }
         diff == 0
     }
